+++ semcore/data-table/src/components/DataTable/DataTable.tsx
@@ -359,7 +359,8 @@ class DataTableRoot<D extends DataTableData> extends Component<
   };
 
   hasFocusableInHeader = () => {
-    return this.headerRef.current && hasFocusableIn(this.headerRef.current);
+    return (this.headerRef.current && hasFocusableIn(this.headerRef.current)) ||
+      this.columns.some((column) => column.sortable);
   };
 
   onExpandRow = (expandedRow: DTRow) => {
@@ -652,7 +653,7 @@ class DataTableRoot<D extends DataTableData> extends Component<
     }
 
     let scrollDirection: 'both' | 'horizontal' | 'vertical' | undefined = undefined;
-    const hasWidthSettings = (Boolean(w) && w !== '100%') || Boolean(wMax);
+    const hasWidthSettings = Boolean(w) || Boolean(wMax);
     const hasHeightSettings = (Boolean(h) && h !== 'fit-content') || Boolean(hMax);
 
     if (hasWidthSettings && !hasHeightSettings) {