+++ semcore/dropdown/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.0.3] - 2025-06-16
+## [16.0.3] - 2025-06-20
+
+### Added
+
+- `sticky` property for Dropdown.Group.
 
 ### Change